Revert rename api->api_c
