+++ libvcx/src/api/connection.rs
@@ -1205,7 +1205,7 @@ mod tests {
     use std::ffi::CString;
     use std::ptr;
 
-    use api_c::{return_types_u32, VcxStateType};
+    use api::{return_types_u32, VcxStateType};
     use connection::tests::{build_test_connection_inviter_invited, build_test_connection_inviter_null, build_test_connection_inviter_requested};
     use utils::constants::{DELETE_CONNECTION_DECRYPTED_RESPONSE, GET_MESSAGES_DECRYPTED_RESPONSE};
     use utils::devsetup::*;
