+++ libvcx/src/api/credential.rs
@@ -817,8 +817,8 @@ mod tests {
     use serde_json::Value;
 
     use ::credential::tests::BAD_CREDENTIAL_OFFER;
-    use api_c::return_types_u32;
-    use api_c::VcxStateType;
+    use api::return_types_u32;
+    use api::VcxStateType;
     use utils::constants::{GET_MESSAGES_DECRYPTED_RESPONSE, V3_OBJECT_SERIALIZE_VERSION};
     
     use utils::devsetup::*;
