+++ libvcx/src/api/logger.rs
