+++ libvcx/src/api/return_types_u32.rs
