+++ libvcx/src/credential.rs
@@ -213,7 +213,7 @@ pub fn get_credential_status(handle: u32) -> VcxResult<u32> {
 
 #[cfg(test)]
 pub mod tests {
-    use api_c::VcxStateType;
+    use api::VcxStateType;
     use aries::messages::issuance::credential::Credential as CredentialV3;
     use connection;
     use credential_request::CredentialRequest;
