+++ libvcx/src/schema.rs
@@ -2,7 +2,7 @@ use std::string::ToString;
 
 use serde_json;
 
-use api_c::PublicEntityStateType;
+use api::PublicEntityStateType;
 use error::prelude::*;
 use messages::ObjectWithVersion;
 use utils::constants::DEFAULT_SERIALIZE_VERSION;
